% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{cytolib_LdFlags}
\alias{cytolib_LdFlags}
\title{Output the LD flags for building against cytolib. These flags are propagated
to sourceCpp via the inlineCxxPlugin (defined below) and to packages 
via a line in Makevars[.win] like this:}
\usage{
cytolib_LdFlags()
}
\description{
PKG_LIBS += $(shell "${R_HOME}/bin${R_ARCH_BIN}/Rscript.exe" -e "cytolib::cytolib_LdFlags()")
}
