% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{LdFlags}
\alias{LdFlags}
\title{linker flags
Generate linker flags
This function prints the absolute path to the protobuf library file that is shipped with the package.
So that user package can use it in their Makevar file.}
\usage{
LdFlags()
}
\value{
nothing.
}
\description{
linker flags
Generate linker flags
This function prints the absolute path to the protobuf library file that is shipped with the package.
So that user package can use it in their Makevar file.
}
\examples{
LdFlags()
}
